+++ src/Resources/contao/config/config.php
@@ -4,3 +4,4 @@
  * Hooks
  */
 $GLOBALS['TL_HOOKS']['parseTemplate']['speed'] = ['huh.speed.listener.hooks', 'parseTemplate'];
+$GLOBALS['TL_HOOKS']['addImageToTemplateData']['speed'] = ['huh.speed.listener.hooks', 'addImageToTemplateData'];