+++ tests/ContaoManager/PluginTest.php
@@ -41,7 +41,7 @@ class PluginTest extends TestCase
 
         static::assertCount(1, $bundles);
         static::assertInstanceOf(BundleConfig::class, $bundles[0]);
-        static::assertEquals(HeimrichHannotContaoSpeedBundle::class, $bundles[0]->getName());
-        static::assertEquals([ContaoCoreBundle::class], $bundles[0]->getLoadAfter());
+        static::assertSame(HeimrichHannotContaoSpeedBundle::class, $bundles[0]->getName());
+        static::assertSame([ContaoCoreBundle::class], $bundles[0]->getLoadAfter());
     }
 }